requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Flask-Migrate==4.0.5
Werkzeug==3.1.3
Jinja2==3.1.5
SQLAlchemy==2.0.36
python-dotenv==1.0.1
markdown==3.7
pytest==8.3.4
gunicorn==22.0.0
psycopg[binary]>=3.1.0
docker==7.0.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
